Fix: Apply float32 dtype to search query embeddings
Same FAISS dtype error was happening in search() method.
Fixed create_index() but missed search() - now both are float32.

Error: 'in method fvec_renorm_L2, argument 3 of type float *'
Location: search.py line 183 (query normalization)
Fix: Add dtype=np.float32 to query_embedding array

+++ ine_mcp_python/ine_mcp/embeddings/search.py
@@ -177,13 +177,14 @@ class SemanticSearchEngine:
 
         # Encode query using FastEmbed
         query_embeddings = list(self.model.embed([query]))
-        query_embedding = np.array([query_embeddings[0]])  # Shape: (1, dim)
+        # CRITICAL: Must be float32 for FAISS compatibility
+        query_embedding = np.array([query_embeddings[0]], dtype=np.float32)  # Shape: (1, dim)
 
-        # Normalize for cosine similarity
+        # Normalize for cosine similarity (requires float32)
         faiss.normalize_L2(query_embedding)
 
-        # Search
-        scores, indices = self._index.search(query_embedding.astype(np.float32), top_k)
+        # Search (already float32, no conversion needed)
+        scores, indices = self._index.search(query_embedding, top_k)
 
         # Filter by threshold and format results
         results = []
